+++ packages/terra-action-header/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Changed
   * Updated WDIO screenshots due to webpack-config-terra sass upgrade.
+  
+## 2.71.0 - (August 25, 2021)
+
+* Changed
+  * Updated WDIO snapshots due to terra-functional-testing v2 upgrade.
 
 ## 2.70.0 - (July 6, 2021)