+++ packages/terra-form-textarea/CHANGELOG.md
@@ -5,6 +5,11 @@
 * Changed
   * Updated WDIO Screenshots due to webpack-config-terra sass upgrade.
 
+## 5.18.0 - (August 25, 2021)
+
+* Changed
+  * Updated WDIO snapshots due to terra-functional-testing v2 upgrade.
+
 ## 5.17.0 - (June 22, 2021)
 
 * Changed