+++ packages/terra-paginator/CHANGELOG.md
@@ -5,6 +5,14 @@
 * Changed
   * Updated WDIO Screenshots due to webpack-config-terra sass upgrade.
 
+## 2.76.0 - (August 25, 2021)
+
+* Added
+  * Added prop `pageLabel` to customize page label.
+
+* Changed
+  * Updated WDIO snapshots due to terra-functional-testing v2 upgrade.
+
 ## 2.75.0 - (July 6, 2021)
 
 * Changed