+++ packages/terra-table/CHANGELOG.md
@@ -5,6 +5,11 @@
 * Changed
   * Updated WDIO Screenshots due to webpack-config-terra sass upgrade.
 
+## 4.28.0 - (August 25, 2021)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.27.0 - (July 6, 2021)
 
 * Changed